requirements.txt
anthropic>=0.18.0
pdfplumber>=0.10.0
python-docx>=1.1.0
PyPDF2>=3.0.0
tkinterdnd2>=0.3.0